conf: Update WINDRES and RC variables
We need to ensure that WINDRES searches the correct include directory
for header files. Now that a default sysroot is specified to an invalid
location, it highlights that we need to specify the location manually and
explictly.

autotools appears to use "RC" in preference to "WINDRES" so we need
to export that to ensure it doesn't just find the utility and use it
without an include directory.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>