*-mingw32.conf: Add whitelist of WINDRES and RC to hash
WINDRES and RC values are exported when building for a mingw32 target.  These
values, due to being exported, end up in the signature hash for everything in
the system.  By adding these to the whitelist we prevent contamination and
allow the system to properly reuse existing -native and other sstate-cache.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>