gettext_0.19.%.bbappend: Handle packaging .dll's for mingw32
Package the libgettext*.dll files in their respective packages.

Additionally package libintl*.dll in a separate libintl package, this
avoids the need to ship libintl*.dll in gettext-runtime which pulls in a
number of files and binaries which are not required for libintl on its
own. Whilst the package is added to PACKAGES unconditionally, it remains
empty for targets where libintl is not provided/built by gettext.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>