libsdl: Windows specific PACKAGECONFIG
Add a PACKAGECONFIG for 'stdio-redirect'. This configure option
enables/disables the Windows only option for redirecting the
stdout/stderr to files instead of to the console.

Add the 'directx' PACKAGECONFIG to configure whether to build with
DirectX support. By default this is enabled (when targeting
mingw32/windows) however this requires the DirectX libraries in order to
include and link against.

Disable X11 support when building for the nativesdk mingw32.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>