Fix Yocto Project compatible errors
Each of the recipes modified fail the YP compatible checking script because
they modify the checksum of the package, even when the mingw32 SDK is not
enabled.

This is corrected by ensuring that each of the settings is only done when
the mingw32 override is enabled.

The PACKAGECONFIG options are moved to oe-core, as there is no way to
specify they should only available in the mingw32 override is enabled.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>