classes/toolchain-scripts-mingw32: Fix bad cmd flag
The command to properly set the exit code at the end of the toolchain
environment was using the correct flag delimiter '/', but the code to
coerce all the unix-style paths to windows paths was incorrectly
changing this to a '\'. This worked find in wines cmd, but on an actual
Windows host, this caused cmd to wait indefinitely. Fix this by
appending the command after the path separators are coerced.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>