gcc/binutils: Allow plugins to be enabled on cross-canadian toolchain
As of GCC 8, plugin support is available with mingw.  If the plugins are
enabled in the regular binutils/gcc configurations, carry over that
configuration with the mingw versions as well.

When plugins are enabled, they produce a file called [plugin].dll.a.  This
causes the debugsrc processor to find the file and attempt to pull out dwarf
information.  However, we need to prevent this as the files are not ELF, but
in Microsoft PE format, and trigger an error if processed.

Signed-off-by: Mark Hatle <mark.hatle@kernel.crashing.org>
Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>