Switch to HOSTTOOLS_NONFATAL
The changes to split classes into global vs. image specific contexts
has broken the inclusion of `wine` and `wineserver` host tools for
testing MinGW SDKs. This is because testsdk is an image specific class
and therefore it's inclusion is not detected globally and the wine host
tools are not present so the SDK tests fail.

Resolve this by using HOSTTOOLS_NONFATAL which will include the tools if
they exist, but won't fail if they are not present. This does mean that
users will now not know they need wine "up front" when doing a build,
but it will instead fail later when they actually try to test the SDK,
but there isn't really a better way to fix this.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>